py/requirements.txt
PyICU
pyyaml
taglib